Initial Commit
This has most of the functionality. The biggest issue is, however, the
host is using WCF, which I have learned through extensive googling, is
not supported in .NET core at this time.

This is only slightly vexing, but can be mitigated via some other web
hosting framework like Nancy. I'm looking into implementing Nancy as the
host currently, but I need this off my disk for now.