Use of Nancy
Initial use of Nancy.
I'm not terribly proud of the URL schema.  Hard coding it all the  way
down is not my favourite idea.  I'm fairly sure I can get  this to work
semi generically with some Expression Trees, but I'm not well-versed in
them enough to work them out  at this time.